+++ CHANGELOG.md
@@ -1,3 +1,37 @@
+## [Release 2.8.4](https://github.com/aws-amplify/amplify-android/releases/tag/release_v2.8.4)
+
+### Bug Fixes
+- **datastore:** send only distinct network events after subscribe ([#2451](https://github.com/aws-amplify/amplify-android/issues/2451))
+
+[See all changes between 2.8.3 and 2.8.4](https://github.com/aws-amplify/amplify-android/compare/release_v2.8.3...release_v2.8.4)
+
+## [Release 2.8.3](https://github.com/aws-amplify/amplify-android/releases/tag/release_v2.8.3)
+
+### Bug Fixes
+- **datastore:** SerializedCustomType de/serialization ([#2350](https://github.com/aws-amplify/amplify-android/issues/2350))
+- **auth:** Use java.time.Instant for AWSTemporaryCredentials expiration ([#2415](https://github.com/aws-amplify/amplify-android/issues/2415))
+- **datastore:** Load pending mutations one at a time from outbox ([#2428](https://github.com/aws-amplify/amplify-android/issues/2428))
+- **datastore:** register network callback only once in reachability monitor ([#2434](https://github.com/aws-amplify/amplify-android/issues/2434))
+- **api:** Allow null associations on non-required fields. ([#2440](https://github.com/aws-amplify/amplify-android/issues/2440))
+
+[See all changes between 2.8.2 and 2.8.3](https://github.com/aws-amplify/amplify-android/compare/release_v2.8.2...release_v2.8.3)
+
+## [Release 2.8.2](https://github.com/aws-amplify/amplify-android/releases/tag/release_v2.8.2)
+
+### Miscellaneous
+- fix(all) Fix non-existent aws-http version + ensure Smithy version matches one in Kotlin SDK ([#2425](https://github.com/aws-amplify/amplify-android/issues/2425))
+
+[See all changes between 2.8.1 and 2.8.2](https://github.com/aws-amplify/amplify-android/compare/release_v2.8.1...release_v2.8.2)
+## [Release 2.8.1](https://github.com/aws-amplify/amplify-android/releases/tag/release_v2.8.1)
+
+### Bug Fixes
+- String interpolation for workflow ([#2412](https://github.com/aws-amplify/amplify-android/issues/2412))
+- Restore notifying comments on pending-close-response issues ([#2417](https://github.com/aws-amplify/amplify-android/issues/2417))
+
+### Miscellaneous
+- Revert "fix(datastore): Fix MutationProcessorRetryTest" ([#2419](https://github.com/aws-amplify/amplify-android/issues/2419))
+
+[See all changes between 2.8.0 and 2.8.1](https://github.com/aws-amplify/amplify-android/compare/release_v2.8.0...release_v2.8.1)
 ## [Release 2.8.0](https://github.com/aws-amplify/amplify-android/releases/tag/release_v2.8.0)
 
 ### Features